refactor: DTO 클래스 위치 이동 및 스케줄 테이블 volunteerNum 필드 누락에 따른 추가
