+++ src/test/java/project/volunteer/domain/recruitment/application/RecruitmentDtoServiceImplTest.java
@@ -19,7 +19,7 @@ import project.volunteer.domain.image.domain.ImageType;
 import project.volunteer.domain.image.domain.RealWorkCode;
 import project.volunteer.domain.participation.dao.ParticipantRepository;
 import project.volunteer.domain.participation.domain.Participant;
-import project.volunteer.domain.recruitment.application.dto.ParticipantState;
+import project.volunteer.global.common.response.ParticipantState;
 import project.volunteer.domain.recruitment.application.dto.RecruitmentDetails;
 import project.volunteer.domain.recruitment.application.dto.RecruitmentParam;
 import project.volunteer.domain.recruitment.dao.RecruitmentRepository;
